initial commit CountOfNumbers.java

+++ src/com/java/fundamentals/CountOfNumbers.java
@@ -6,11 +6,12 @@ public class CountOfNumbers {
     static void main() {
         Scanner sc = new Scanner(System.in);
         System.out.println("enter the value:");
-        int N = sc.nextInt();
+        long N = sc.nextLong();
         int Count = 0;
         while(N !=0) {
             N = N / 10;
             Count++;
+
         }
         System.out.println("print count:" + Count);
     }
